backend/test/unit/modules/lc/assets/destroy_asset.spec.ts
import { DestroyAsset } from '../../../../../src/app/modules/lc/assets/destroy_asset';

describe('DestroyAsset', () => {
  let transactionAsset: DestroyAsset;

	beforeEach(() => {
		transactionAsset = new DestroyAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(3);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('destroy');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('apply', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
